requirements.txt
joblib==1.2.0
numpy==1.23.5
pandas==2.0.2
PyYAML==6.0
scikit-learn==1.2.2
xgboost==2.0.0